EFI/Firmware: 'svn merge ^/vendor/edk2/UDK2010.SR1 ^/vendor/edk2/current .', reverting and removing files+dirs listed in ReadMe.vbox, resolving conflicts with help from ../UDK2014.SP1/. This is a raw untested merge.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@58459 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/EFI/Firmware/StdLib/BsdSocketLib/res_debug.c
@@ -243,7 +243,7 @@ fp_nquery(const u_char *msg, int len, FILE *file) {
     if ((!_res.pfcode) || (_res.pfcode & RES_PRF_HEADX) || rcode)
         fprintf(file,
             ";; ->>HEADER<<- opcode: %s, status: %s, id: %d\n",
-            _res_opcodes[opcode], _res_resultcodes[rcode], id);
+            _res_opcodes[opcode], _res_resultcodes[rcode], (int)id);
     if ((!_res.pfcode) || (_res.pfcode & RES_PRF_HEADX))
         putc(';', file);
     if ((!_res.pfcode) || (_res.pfcode & RES_PRF_HEAD2)) {
